Remove rudimentary metrics collection from Sup
* Remove dependency on prometheus in Supervisor crate
* Remove "/metrics" endpoint from Supervisor's http gateway

The metrics endpoint was added before the vision of the Supervisor
was fully realized. It hasn't been in it's current state provides
simple counters for how many times an endpoint was called.

We should revisit metrics within the Supervisor at a later date
when a set of fresh eyes if we want to provide some sort of more
useful data collection (such as size of rumor store or other garbage
collection related statistics)

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>